requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
accelerate==0.27.2
transformers==4.38.2
-e git+https://github.com/huggingface/trl.git@4219cbfedcf27ef75754f87079c0d97b767f27f0#egg=trl
datasets==2.17.1
mlflow==2.10.2
peft==0.9.0
torch==2.1.0+cu118
boto3==1.34.45
urllib3==1.26.6
numpy==1.26.4
sentencepiece==0.2.0
openai==1.16.1
python-dotenv==1.0.1
adjustText==1.1.1
